requirements.txt
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
folium==0.12.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
PyJWT==1.7.1
pyngrok==5.0.5
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
twilio==6.55.0
urllib3==1.26.4
visitor==0.1.3
Werkzeug==1.0.1
